third_party/WebKit/LayoutTests/shadow-dom/imperative-api-notslotchangefornotassignfunction.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
    const test = async_test('slotchange event must not be fired');
    let eventCount = 0;
    let anothereventCount = 0;
    let theothereventCount = 0;
    let host;
    let slot1;
    let slot2;
    let slot3;
    let child1;
    let child2;
    let child3;
    let shadowRoot;

    test.step(function () {
        host = document.createElement('div');
        document.body.appendChild(host);

        shadowRoot = host.attachShadow({mode: 'open', slotting: 'manual'});
        slot1 = document.createElement('slot');

        slot1.addEventListener('slotchange', function (event) {
            eventCount++;
        });

        slot2 = document.createElement('slot');

        slot2.addEventListener('slotchange', function (event) {
            anothereventCount++;
        });

        slot3 = document.createElement('slot');

        slot3.addEventListener('slotchange', function (event) {
            theothereventCount++;
        });

        child1 = document.createElement('div');
        child2 = document.createElement('div');

        host.appendChild(child1);

        slot1.assign([child1]);
        slot2.assign([child1]);
        slot3.assign([child2]);
        shadowRoot.appendChild(slot1);
        shadowRoot.appendChild(slot2);

        assert_equals(eventCount, 0, 'slotchange event must not be fired synchronously');
        assert_equals(anothereventCount, 0, 'slotchange event must not be fired synchronously');
        assert_equals(theothereventCount, 0, 'slotchange event must not be fired synchronously');
    });

    setTimeout(function () {
        test.step(function () {
            assert_equals(eventCount, 1, 'slotchange must be fired exactly once after the assigned nodes changed');
            assert_equals(anothereventCount, 0, 'slotchange event must not be fired when slot that have nodes that is assigned another slot is appended.');
            assert_equals(theothereventCount, 0, 'slotchange event must not be fired when slot that have nodes that is not host\'s child is appended.');
            slot2.remove();
            slot1.assign([child2]);
            slot3.remove();
        });

        setTimeout(function () {
            test.step(function () {
                assert_equals(eventCount, 2, 'slotchange must be fired exactly once after the assigned nodes changed');
                assert_equals(anothereventCount, 0, 'slotchange event must not be fired when slot that have nodes that is assigned another slot is removed.');
                assert_equals(theothereventCount, 0, 'slotchange event must not be fired when slot that have nodes that is not host\'s child is removed.');

                slot2.assign([]);
                shadowRoot.appendChild(slot2);
                shadowRoot.appendChild(slot3);
                host.appendChild(child2);
            });

            setTimeout(function () {
                test.step(function () {
                    assert_equals(eventCount, 3, 'slotchange must be fired exactly once after the assigned nodes changed');
                    assert_equals(anothereventCount, 0, 'slotchange event must not be fired when slot that has no child is appended');
                    assert_equals(theothereventCount, 0, 'slotchange event must not be fired when child that assigned another slot is appended');

                    child2.remove()
                });

                setTimeout(function () {
                    test.step(function () {
                        assert_equals(eventCount, 4, 'slotchange must be fired exactly once after the assigned nodes changed');
                        assert_equals(theothereventCount, 0, 'slotchange event must not be fired when child that is assigned another slot is deleted');
                    });
                    test.done();
                }, 1);
            }, 1);
        }, 1);
    }, 1);
</script>
</body>
</html>
